requirements.txt
alembic==1.7.5
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
configparser==5.1.0
crayons==0.4.0
cryptography==35.0.0
dependency-injector==4.37.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
injector==0.18.4
kafka-python==2.0.2
Mako==1.1.5
MarkupSafe==2.0.1
outcome==1.1.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
PyMySQL==1.0.2
pyOpenSSL==21.0.0
python-dotenv==0.19.2
requests==2.26.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.27
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
webdriver-manager==3.5.2
wsproto==1.0.0
zipp==3.6.0
